Publishing run loop info to a topic and saving all of the robot state information into robot_state_data.txt

+++ iam_robolib/include/iam_robolib/trajectory_generator/counter_trajectory_generator.h
@@ -17,15 +17,15 @@ class CounterTrajectoryGenerator : public TrajectoryGenerator {
 
   void get_next_step() override;
 
-  float current_point_[3] = {};
-  float delta_ = 0;
+  double current_point_[3] = {};
+  double delta_ = 0;
 
  private:
   int start_ = 0;
   int end_ = 0;
   int current_val_ = 0;
 
-  float start_point_[3] = {};
+  double start_point_[3] = {};
 };
 
 #endif  // IAM_ROBOLIB_TRAJECTORY_GENERATOR_COUNTER_TRAJECTORY_GENERATOR_H_
